Fix StringIndexOutOfBoundsException
In some locales (e.g. Chinese), the method getShortWeekdays() in
 ateFormatSymbols used in the CaldarView class will return an array
of string which are only two characters long.

In this case, String.substring() method in the loop, which was
hardcoded with parameter (0, 3) will throw a StringIndexOutOfBoundsException
since there is only two characters in the returned string.

This commit fixes this behavior by first determining the length of the
string, then use either the length of the string or 3 as the second
parameter to the substring() method.